feat: adding converstion factors
