+++ cc-mage/data_exporters/load_conversion_factors_db.py
@@ -17,7 +17,7 @@ def export_data_to_postgres(df: DataFrame, **kwargs) -> None:
     Docs: https://docs.mage.ai/design/data-loading#postgresql
     """
     schema_name = 'raw_data'  # Specify the name of the schema to export data to
-    table_name = 'ipcc_energy_density'  # Specify the name of the table to export data to
+    table_name = 'ipcc_conversion_factor'  # Specify the name of the table to export data to
     config_path = path.join(get_repo_path(), 'io_config.yaml')
     config_profile = 'default'
